pyproject.toml
[project]
name = "agent-mesh-pipeline"
version = "0.1.0"
description = "Feed, signal, and risk workers for Agent Mesh"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "redis>=5.0.0",
  "httpx>=0.27.0",
  "pydantic>=2.6.0",
  "websocket-client>=1.8.0",
]

[project.optional-dependencies]
dev = ["ruff>=0.4.0", "pytest>=8.0.0"]

[tool.ruff]
line-length = 100

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_mesh_pipeline"]
